+++ database/migrations/2024_04_24_211823_create_logros_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('description');
-            $table->integer('point');
+            $table->decimal('point', 64, 0);
             $table->timestamps();
         });
     }
